# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mini Sentry API
  version: 0.7.0
servers:
  - url: http://localhost:8000
paths:
...
  /api/events/{id}/:
    get:
      summary: Retrieve event
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
...
  /api/events/clickhouse:
    get:
      summary: Query ClickHouse events
      parameters:
        - in: query
          name: project
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClickHouseEvent'
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id: { type: integer }
        project: { type: integer }
        group: { type: integer, nullable: true }
        message: { type: string }
        level: { type: string, enum: [error, warning, info] }
        payload: { type: object }
        received_at: { type: string, format: date-time }
        release: { type: integer, nullable: true }
        environment: { type: string }
        stack: { type: string, nullable: true }
        symbolicated: { type: object }
...
    ClickHouseEvent:
      type: object
      properties:
        id: { type: integer }
        project: { type: string }
        level: { type: string }
        fingerprint: { type: string }
        title: { type: string }
        message: { type: string }
        received_at: { type: string }
...